added subject_agnostic_component
